+++ src/main/kotlin/org/github/telegabots/entity/CommandBlock.kt
@@ -7,10 +7,10 @@ import java.time.LocalDateTime
  * Entity related with message and user
  */
 data class CommandBlock(val messageId: Int,
-                        val userId: Int,
+                        val userId: Long,
                         val messageType: MessageType,
                         val id: Long = 0L,
                         val createdAt: LocalDateTime = LocalDateTime.now()
 ) {
-    fun isValid(): Boolean = messageId != 0 && userId != 0
+    fun isValid(): Boolean = messageId != 0 && userId != 0L
 }
